+++ src/lib/index.js
@@ -3,5 +3,6 @@ import Search from "./search.svelte";
 import Filter from "./filter.svelte";
 import WebinarOverview from "./webinar-overview.svelte";
 import Navigation from "./navigation.svelte";
+import QandA from "./q-a.svelte";
 
-export {ContouringOverview, Search, Filter, WebinarOverview, Navigation}
+export {ContouringOverview, Search, Filter, WebinarOverview, Navigation, QandA}
