+++ src/routes/webinars/[slug]/+page.svelte
@@ -1,19 +1,6 @@
 <script>
-  import { postComment } from '$lib/api.js';
-
+  import { QandA } from "$lib/index.js";
   export let data; 
-
-  let content = ''
-  let webinarId = data.webinar.id;
-
-  async function handleSubmit() {
-    try {
-      await postComment(content, webinarId);
-      content = ''; // Clear the input field on success
-    } catch (error) {
-      console.error('Error posting comment:', error);
-    }
-  }
 </script>
 
 <main>
@@ -29,22 +16,9 @@
   </div>
   
   <section>
-    <h3>Q&A</h3>
-  
-    <form on:submit|preventDefault={handleSubmit}>
-      <label>Ask a question.
-        <input name="comment" placeholder="Add a comment..." bind:value={content}>
-      </label>
-      <button type="submit">Send</button>
-    </form>
-  
-    <section class="comments">
-      <h4>Comments</h4>
-  
-      {#each data.comments as comment}
-        <p>{comment.content}</p>
-      {/each}
-    </section>
+    <QandA 
+      webinarId = {data.webinar.id}
+      comments = {data.comments} />
   </section>
 </main>
